requirements.txt
langchain==1.2.3
langchain-core==1.2.7
langchain-groq==1.1.1
numpy==2.4.1
pandas==2.3.3
plotly==6.5.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
ruamel.yaml==0.19.1
streamlit==1.52.2
yfinance==1.0
